+++ src/card/cards.rs
@@ -29,6 +29,7 @@ mod regeneration_remaining;
 mod spore;
 mod tapped_and_untap;
 mod targeted_regeneration;
+mod unleash;
 pub use activation_timing::*;
 pub use attack_restrictions::*;
 pub use battalion::*;
@@ -56,6 +57,7 @@ pub use regeneration_remaining::*;
 pub use spore::*;
 pub use tapped_and_untap::*;
 pub use targeted_regeneration::*;
+pub use unleash::*;
 
 pub const ANKH_OF_MISHRA: CardDefinitionId = CardDefinitionId(1);
 pub const ATOG: CardDefinitionId = CardDefinitionId(2);